+++ src/TransactionsContext.tsx
@@ -1,0 +1,35 @@
+import { createContext, ReactNode, useEffect, useState } from 'react';
+import api from './services/api';
+
+
+interface Transaction {
+  id: number;
+  title: string;
+  amount: number;
+  type: string;
+  category: string;
+  createdAt: string;
+}
+
+interface TransactionProviderProps {
+  children: ReactNode;
+}
+
+export const TransactionsContext = createContext<Transaction[]>([]);
+
+export function TransactionsProvider({ children }: TransactionProviderProps) {
+  const [transactions, setTransactions] = useState<Transaction[]>([]);
+
+  useEffect(() => {
+    api.get('/transactions')
+      .then(res => {
+        setTransactions(res.data.transactions)
+      })
+  },[])
+
+  return (
+    <TransactionsContext.Provider value={transactions}>
+      { children }
+    </TransactionsContext.Provider>
+  )
+}
