+++ src/components/Header/index.tsx
@@ -1,6 +1,7 @@
-import React from 'react';
+import React, { useContext } from 'react';
 
 import logoImg from '../../assets/logo.svg';
+import { TransactionsContext } from '../../TransactionsContext';
 import { Container, Content } from './styles';
 
 interface HeaderProps {
@@ -8,7 +9,9 @@ interface HeaderProps {
 }
 
 export function Header({ onOpenNewTransacionModal }: HeaderProps) {
+  const transactions = useContext(TransactionsContext);
 
+  console.log('header ' + JSON.stringify(transactions));
   return (
     <Container>
       <Content>
