+++ src/components/TransactionsTable/index.tsx
@@ -1,27 +1,10 @@
-import React, { useEffect, useState } from 'react'
-import api from '../../services/api'
+import React, { useContext } from 'react'
+import { TransactionsContext } from '../../TransactionsContext'
 import { Container } from './styles'
-
-interface Transaction {
-  id: number;
-  title: string;
-  amount: number;
-  type: string;
-  category: string;
-  createdAt: string;
-}
 
 export function TransactionsTable() {
 
-  const [transactions, setTransactions] = useState<Transaction[]>([]);
-
-  useEffect(() => {
-    api.get('/transactions')
-      .then(res => {
-        setTransactions(res.data.transactions)
-      })
-  },[])
-
+  const transactions = useContext(TransactionsContext);
   return (
     <Container>
       <table>
